// src/interfaces/common/category.interface.ts
export interface Category {
  _id?: string;
  readOnly?: boolean;
  name?: string;
  nameBn?: string;
  nameIt?: string;
  totalPost?: string;
  slug?: string;
  image?: string;
  mobileImage?: string;
  description?: string;
  serial?: number;
  status?: string;
  postType?: string;
  select?: boolean;
}